+++ crates/px-fits/src/lib.rs
@@ -1,0 +1,5 @@
+// Soon to be wrapper around https://github.com/cds-astro/fitsrs
+// * Lazy data loading
+// * pure rust
+// * doesn't write images well but that's ok
+// * can support wasm if all data loaded outside of wasm (fetch)