+++ crates/px/src/lib.rs
@@ -22,8 +22,6 @@ pub async fn run(cli: Cli) -> Result<ExitStatus> {
     // Configure the `Printer`, which controls user-facing output in the CLI.
     let printer = Printer::Default;
 
-    let file = "'/Users/kyle/Pictures/Astro/FRA300_6200mc_pro/LIGHT/NGC 2244/2026-02-02/RAW_O_300.00s/2026-02-02_22-49-25_O_-9.90c_100g_30o_300.00s_d_1x1_0720.fits'";
-
     match cli.command {
         Commands::GenerateShellCompletion(args) => {
             let mut cmd = Cli::command();
@@ -31,8 +29,6 @@ pub async fn run(cli: Cli) -> Result<ExitStatus> {
             clap_complete::generate(args.shell, &mut cmd, &bin_name, &mut stdout());
             Ok(ExitStatus::Success)
         }
-
-        Commands::Check => commands::stat(file, printer).await,
 
         Commands::SirilTest => commands::siril_test(printer).await,
 
@@ -63,5 +59,7 @@ pub async fn run(cli: Cli) -> Result<ExitStatus> {
         Commands::Profile(_profile_namespace) => todo!(),
 
         Commands::Tui => commands::terminal_ui(printer).await,
+
+        Commands::Inspect(args) => commands::inspect_file(args, printer).await,
     }
 }