+++ crates/siril-sys/src/siril_ext.rs
@@ -1,3 +1,5 @@
-/// Prelude file for all Siril Extensions from Commands
-/// import with `use siril_sys::siril_ext::*;`
+//! Prelude file for all Siril Extensions from Commands
+//! import with `use siril_sys::siril_ext::*;`
+
+pub use crate::commands::load::LoadExt;
 pub use crate::commands::pwd::PwdExt;