+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Add shell completion, use `px generate-shell-completion zsh` to generate
 * Updated README.md with install, completiong and uninstall instructions
 * Add an `OutputSink` to siril builder to redirect stdout/stderr
+* Removed `stat` command and replaced with `inspect`
+* Turned off logging by default and moved to `-v, -vv, -vvv` global cli options
 
 ## 0.1.3